requirements.txt
zarr
pynrrd
tifffile
scikit-image
